Add more optimized pubsub
