+++ indexer/httpd/src/error.rs
@@ -1,4 +1,4 @@
-use {sea_orm::sqlx, std::io, thiserror::Error};
+use {indexer_sql::error::IndexerError, sea_orm::sqlx, std::io, thiserror::Error};
 
 #[derive(Error, Debug)]
 pub enum Error {
@@ -10,4 +10,7 @@ pub enum Error {
 
     #[error(transparent)]
     Sqlx(#[from] sqlx::Error),
+
+    #[error(transparent)]
+    Indexer(#[from] IndexerError),
 }
